First graphics functions
Add isTouched,.
Init tft in constructor. Add clearScreen, fillScreen.
A list of the other MindstormsUI functions are in the header. Many of
them are extraneous. For the others you might as well just call that
function yourself (on tft from the Adafruit GFX library). tft is private
right now, so make it public or add wrapper methods.
Also note the Adafruit_ILI9341 automatically draws, no need to refresh
or use a display parameter.