First println utilities
From the user's perspective nothing will change, they still simply call
.println on the MindsensorsUI object, the screen member of the EVShield
instance. However, now in the constructor we increase the text size (the
default, 1, is very small, so double the x and y dimentions) and set the
text color (by default it is white text with a clear background,
explicitly specify black for opaque). This also overrides println to
first check if the cursor is off the bottom of the screen, wrapping it
to the top if so. This is a simple change but will not work if the user
tries to print a single very long string that wraps past the bottom edge
of the screen.