Fix LED patterns
The previous impementations of these methods did not work because it
reset breathNow to 0 every time it was called. Instead declare and
initialize the static timer variable outside of the function. Use delay
instead of delayMicroseconds because the latter does not let the ESP8266
run background WiFi and TCP/IP tasks. I could not tell what
ledHeartBeatPattern was supposed to be, it just seemed like a more
choppy ledBreathingPattern. Now it shows two quick beats then a pause.