Improve anti-cheat and LED flashing
Tapping the button will now also be detected as cheating and reset the
timer to a new random value. You should be able to press the button at
any time to continue now, not just between LED flashing cycles. Also, if
an I2C call fails to set the LED color, it will be attempted again in
50ms and the error will be less noticable.